+++ frontend/src/routes/__root.tsx
@@ -1,7 +1,12 @@
 // import { TanStackRouterDevtools } from "@tanstack/router-devtools";
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
-import { createRootRoute, Outlet } from "@tanstack/react-router";
+import {
+  createRootRoute,
+  ErrorComponent,
+  Outlet,
+} from "@tanstack/react-router";
+import posthog from "@/posthog";
 import Snackbar from "../components/Snackbar/Snackbar";
 import Header from "../features/navigation/components/Header";
 
@@ -22,4 +27,8 @@ function C() {
 
 export const Route = createRootRoute({
   component: () => <C />,
+  errorComponent: ({ error }) => {
+    posthog.captureException(error);
+    return <ErrorComponent error={error} />;
+  },
 });